+++ ent/meeting/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Meeting {
+func ID(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Meeting {
+func IDEQ(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Meeting {
+func IDNEQ(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Meeting {
+func IDIn(ids ...int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Meeting {
+func IDNotIn(ids ...int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Meeting {
+func IDGT(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Meeting {
+func IDGTE(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Meeting {
+func IDLT(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Meeting {
+func IDLTE(id int64) predicate.Meeting {
 	return predicate.Meeting(sql.FieldLTE(FieldID, id))
 }
